pyproject.toml
[tool.poetry]
name = "virtualassistant"
version = "0.1.0"
description = "Virtual assistant to compile simple analysis"
authors = ["Cadorini95 <mcadorini@peq.coppe.ufrj.br>"]
readme = "README.md"
packages = [{ include = "virtualassistant" }]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.24"
langchain-openai = "^0.3.14"
langchain-core = "^0.3.55"
python-dotenv = "^1.1.0"
streamlit = "^1.44.1"
sqlalchemy = "^2.0.40"
pyodbc = "^5.2.0"
langchain-community = "^0.3.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
